+++ pysui/sui/sui_pgql/pgql_schema.py
@@ -147,22 +147,40 @@ def load_schema_cache(
             fetch_schema_from_transport=True,
         )
 
-        with _iclient as session:
-            # session.fetch_schema()
-            _long_version = session.transport.response_headers[
-                Schema.SCHEMA_HEADER_SCHEMA_KEY
-            ]
-            _base_version = ".".join(_long_version.split(".")[:2])
-            _schema: DSLSchema = DSLSchema(_iclient.schema)
-            qstr, fndeser = pgql_config(genv, _base_version)
-            try:
-                _rpc_config = fndeser(session.execute(gql(qstr)))
-                _rpc_config.gqlEnvironment = genv
-                schema_mgr._cache_schema_set(
-                    _url, _iclient, _base_version, _long_version, _schema, _rpc_config
-                )
-            except texc.TransportQueryError as gte:
-                schema_mgr._present_schemas.remove(_base_version)
-                schema_mgr._all_versions.remove(_base_version)
+        try:
+            with _iclient as session:
+                # session.fetch_schema()
+                _long_version = session.transport.response_headers[
+                    Schema.SCHEMA_HEADER_SCHEMA_KEY
+                ]
+                _base_version = ".".join(_long_version.split(".")[:2])
+                _schema: DSLSchema = DSLSchema(_iclient.schema)
+                qstr, fndeser = pgql_config(genv, _base_version)
+                try:
+                    _rpc_config = fndeser(session.execute(gql(qstr)))
+                    _rpc_config.gqlEnvironment = genv
+                    schema_mgr._cache_schema_set(
+                        _url,
+                        _iclient,
+                        _base_version,
+                        _long_version,
+                        _schema,
+                        _rpc_config,
+                    )
+                except (
+                    httpx.HTTPError,
+                    texc.TransportQueryError,
+                    texc.TransportServerError,
+                ) as gte:
+                    schema_mgr._present_schemas.remove(_base_version)
+                    schema_mgr._all_versions.remove(_base_version)
+
+        except (
+            httpx.HTTPError,
+            texc.TransportQueryError,
+            texc.TransportServerError,
+        ) as gte:
+            schema_mgr._present_schemas.remove(_scm_ver)
+            schema_mgr._all_versions.remove(_scm_ver)
 
     return schema_mgr