+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.68.0] - Unpublished
+
+### Added
+
+- Docs for native types used in transaction builder arguments
+
+### Fixed
+
+### Changed
+
+- Bump version 0.68.0
+
+### Removed
+
 ## [0.67.0] - 2024-08-21
 
 ### Added